+++ native/libffi/testsuite/libffi.call/strlen4.c
@@ -8,9 +8,9 @@
 
 #include "ffitest.h"
 
-static size_t ABI_ATTR my_f(float a, char *s, int i)
+static int ABI_ATTR my_f(float a, char *s, int i)
 {
-  return (size_t) ((int) strlen(s) + (int) a + i);
+  return (int) strlen(s) + (int) a + i;
 }
 
 int main (void)
